// src/app/types.ts
// app/types.ts
export interface ResumeFormData {
    name: string;
    fatherName: string;
    email: string;
    mobile: string;
    address: string;
    dob: string;  // Date of Birth
    maritalStatus: string;
    nationality: string;
    religion: string;
    gender: string;
    nic: string;
    sureName: string;
    education: string;
    projects: string;
    experience: string;
    languages: string;
    skills: string;
  }
  